refactor(cluster): extract MST-split + confidence kernels to pyo3-free cluster-core (A1)
Step A1 of sharing the clustering MST-split kernel with TS (standalone-ts
medium, Family A). The `find` / `mst_split_components` / `severe_bridge_count` /
`cluster_confidence` functions + the `ConfidenceResult` alias move out of the
pyo3 `native` crate into a NEW pyo3-free, std-only `goldenmatch-cluster-core`
crate (mirrors score-core / graph-core: standalone empty-workspace, path dep of
native, excluded from the bridge workspace). native keeps thin `#[pyfunction]`
shims that delegate, so the pymodule exports the same symbols and the Arrow
orchestrators (build_clusters_native/arrow) call the imported core fns directly.

Byte-parity preserved verbatim (float summation order, strict `<` first-min-wins
tie-break, Kruskal stable-sort edge selection). cluster-core ships 10 unit tests
locking those invariants (tied-weakest-edge determinism, confidence formula,
size<=1 -> 1.0, first-min bottleneck). `cargo test` (cluster-core) + `cargo
check` (native) both green. No behavior change; this is the foundation for the
A2 cluster-wasm surface + TS reroute of splitOversizedCluster.
